feat: Add HelixDB Proof-of-Concept integration and E2E tests
- Adds `HELIXDB_EVALUATION.md` and updates `TODO.md`.
- Implements `BaseMemoryBackend` and moves SQLite+FAISS into `memory_legacy.py`.
- Implements `HelixMemoryBackend` and `HelixClient` for AST conversion.
- Adds `ansible/jobs/helixdb.nomad` job file for HelixDB cluster deployment.
- Adds a standalone FastMCP `helix_server.py`.
- Adds `tests/integration/test_helixdb_e2e.py` to validate insertion/retrieval via mocked HelixDB requests.
